Renamed the main package
Renamed Sandman2 to flask_sandman as per the flask extension conventions
