+++ docs/conf.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 #
-# sandman2 documentation build configuration file, created by
+# flask_sandman documentation build configuration file, created by
 # sphinx-quickstart on Fri Oct 24 08:22:11 2014.
 #
 # This file is execfile()d with the current directory set to its
@@ -48,7 +48,7 @@ source_suffix = '.rst'
 master_doc = 'index'
 
 # General information about the project.
-project = u'sandman2'
+project = u'flask_sandman'
 copyright = u'2014, Jeff Knupp'
 
 # The version info for the project you're documenting, acts as replacement for
@@ -201,7 +201,7 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-  ('index', 'sandman2.tex', u'sandman2 Documentation',
+  ('index', 'flask_sandman.tex', u'flask_sandman Documentation',
    u'Jeff Knupp', 'manual'),
 ]
 
@@ -231,7 +231,7 @@ latex_documents = [
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
 man_pages = [
-    ('index', 'sandman2', u'sandman2 Documentation',
+    ('index', 'flask_sandman', u'flask_sandman Documentation',
      [u'Jeff Knupp'], 1)
 ]
 
@@ -245,8 +245,8 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-  ('index', 'sandman2', u'sandman2 Documentation',
-   u'Jeff Knupp', 'sandman2', 'One line description of project.',
+  ('index', 'flask_sandman', u'flask_sandman Documentation',
+   u'Jeff Knupp', 'flask_sandman', 'One line description of project.',
    'Miscellaneous'),
 ]
