+++ examples/example_user_models.py
@@ -1,5 +1,5 @@
 from user_models import User, Blog, Post
-from sandman2 import get_app
+from flask_sandman import get_app
 
 app = get_app('sqlite+pysqlite:///blog.sqlite3', user_models=[User, Blog, Post])
