+++ flask_sandman/__init__.py
@@ -1,0 +1,2 @@
+"""*flask_sandman*'s main module."""
+from flask_sandman.app import get_app, db, AutomapModel
