+++ flask_sandman/templates/admin/index.html
