--- sandman2/__init__.py
@@ -1,2 +1,0 @@
-"""*sandman2*'s main module."""
-from sandman2.app import get_app, db, AutomapModel
